Se actualiza el enrutamiento del proyecto y se modifica la apariencia de los componentes login y signup
