+++ src/views/LoginPage.jsx
@@ -2,6 +2,7 @@ import { useState } from 'react';
 import { Button, Form } from 'react-bootstrap';
 import 'animate.css';
 
+
 export const LoginPage = () => {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
@@ -30,9 +31,10 @@ export const LoginPage = () => {
         setIsLoggedIn(false);
     };
 
+
     if (isLoggedIn) {
         return (
-            <div className='col-md-6 mx-auto animate__animated animate__fadeIn'>
+            <div className='col-md-3 mx-auto animate__animated animate__fadeIn'>
                 <h1 className='mt-3'>Bienvenido!</h1>
 
                 {/* Insertar componente Perfil */}
@@ -45,36 +47,47 @@ export const LoginPage = () => {
     }
 
     return (
-        <div className='col-md-6 mx-auto animate__animated animate__fadeIn'>
-            <h1 className='mt-3'>Login</h1>
 
-            <Form onSubmit={handleSubmit}>
-                <Form.Group className='mb-3' controlId='formBasicEmail'>
-                    <Form.Label>Email</Form.Label>
-                    <Form.Control
-                        type='email'
-                        placeholder='Ingresa tu email'
-                        value={email}
-                        onChange={(e) => setEmail(e.target.value)}
-                    />
-                </Form.Group>
 
-                <Form.Group className='mb-3' controlId='formBasicPassword'>
-                    <Form.Label>Contraseña</Form.Label>
-                    <Form.Control
-                        type='password'
-                        placeholder='Ingresa tu contraseña'
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}
-                    />
-                </Form.Group>
+        <div className='d-flex justify-content-center align-items-center vh-100' >
+            <div className='col-md-3 mx-auto border border-dark rounded p-5 animate__animated animate__fadeIn'>
+                <h1 className='mt-3'>Login</h1>
 
-                <Button variant='primary btn-dark mb-3' type='submit'>
-                    Enviar
-                </Button>
-            </Form>
+                <Form onSubmit={handleSubmit}>
+                    <Form.Group className='mb-3' controlId='formBasicEmail'>
+                        <Form.Label>Email</Form.Label>
+                        <Form.Control
+                            type='email'
+                            placeholder='Ingresa tu email'
+                            value={email}
+                            onChange={(e) => setEmail(e.target.value)}
+                        />
+                    </Form.Group>
 
-            {error ? <p className='error'>{error}</p> : null}
+                    <Form.Group className='mb-3' controlId='formBasicPassword'>
+                        <Form.Label>Contraseña</Form.Label>
+                        <Form.Control
+                            type='password'
+                            placeholder='Ingresa tu contraseña'
+                            value={password}
+                            onChange={(e) => setPassword(e.target.value)}
+                        />
+                    </Form.Group>
+
+                    <div className="d-flex justify-content-center">
+                        <Button variant='primary btn-dark mb-3' type='submit'>
+                            Enviar
+                        </Button>
+
+                    </div>
+                </Form>
+
+                {error ? <p className='error'>{error}</p> : null}
+            </div>
+
+
         </div>
+
+
     );
 };
